fix authentication controller duplication
